bot/requirements.txt
aiogram==3.6.0
requests==2.31.0
python-dotenv==1.0.1
aiohttp==3.9.5
openai==1.76.0
python-docx==1.1.2